Bugfixes, added some response messages when actions are done, added robustness to prevent the application from crashing.

+++ CapgeminiVoting/Controllers/AdminController.cs
@@ -15,13 +15,7 @@ namespace CapgeminiVoting.Controllers
         [HttpGet]
         public ActionResult Index()
         {
-            return Index(string.Empty);
-        }
-
-        public ActionResult Index(string message)
-        {
-            ViewBag.message = message;
-            IList<EventOverviewModel> model = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+            var model = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
             return View(model);
         }
 
@@ -35,19 +29,26 @@ namespace CapgeminiVoting.Controllers
         [HttpGet]
         public ActionResult ModifyEvent(int id)
         {
-            ViewBag.Title = Resources.Modify_event;
-            EventDetailsModel model = AdminBusinessLayer.GetEventById(id);
-
-            if (model == null)
+            if (AdminBusinessLayer.IsEventOwner(id, User.Identity.GetUserId()))
             {
-                return View("CreateEvent");
+                ViewBag.Title = Resources.Modify_event;
+                var model = AdminBusinessLayer.GetEventById(id);
+
+                if (model == null)
+                {
+                    return View("CreateEvent");
+                }
+
+                var sortedQuestions = model.Questions.ToList();
+                sortedQuestions.Sort();
+                model.Questions = sortedQuestions;
+
+                return View("CreateEvent", model);
             }
 
-            var sortedQuestions = model.Questions.ToList();
-            sortedQuestions.Sort();
-            model.Questions = sortedQuestions;
-            
-            return View("CreateEvent", model);
+            ViewBag.ErrorMessage = Resources.Not_your_event;
+            var indexModel = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+            return View("Index", indexModel);
         }
 
         [HttpPost]
@@ -55,23 +56,51 @@ namespace CapgeminiVoting.Controllers
         {
             bool result = false;
 
-            if (@event != null)
+            if (@event == null)
             {
-                var questionList = @event.Questions.ToList();
-                questionList.RemoveAll(question => string.IsNullOrWhiteSpace(question.Question));
-                foreach (var question in @event.Questions)
+                var model = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+                return View("Index", model);
+            }
+
+            var questionList = @event.Questions.ToList();
+            questionList.RemoveAll(question => string.IsNullOrWhiteSpace(question.Question));
+            @event.Questions = questionList;
+
+            if (@event.Questions.Count == 0)
+            {
+                ModelState.Clear();
+                ModelState.AddModelError(String.Empty, Resources.Please_specify_question);
+                ViewBag.Title = Resources.Create_new_event;
+                return View("CreateEvent", @event);
+            }
+
+            foreach (var question in @event.Questions)
+            {
+                var answerList = question.Answers.ToList();
+                answerList.RemoveAll(answer => string.IsNullOrWhiteSpace(answer.Answer));
+                question.Answers = answerList;
+
+                if (question.Answers.Count < 2 && question.QuestionType != 2)
                 {
-                    var answerList = question.Answers.ToList();
-                    answerList.RemoveAll(answer => string.IsNullOrWhiteSpace(answer.Answer));
-                    question.Answers = answerList;
+                    ModelState.Clear();
+                    ModelState.AddModelError(String.Empty, Resources.Please_specify_answer);
+                    ViewBag.Title = Resources.Create_new_event;
+                    return View("CreateEvent", @event);
                 }
             }
 
             result = AdminBusinessLayer.CreateEvent(@event);
 
             if (result)
-                return RedirectToAction("Index");
-            else return RedirectToAction("CreateEvent");
+            {
+                ViewBag.SuccessMessage = Resources.Event_successfully_created;
+                var indexModel = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+                return View("Index", indexModel);
+            }
+            else
+            {
+                return RedirectToAction("CreateEvent");
+            }
         }
 
         [HttpPost]
@@ -79,11 +108,22 @@ namespace CapgeminiVoting.Controllers
         {
             if (@event == null)
             {
-                RedirectToAction("Index");
+                var model = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+                return View("Index", model);
             }
 
             var questionList = @event.Questions.ToList();
             questionList.RemoveAll(question => string.IsNullOrWhiteSpace(question.Question));
+            @event.Questions = questionList;
+
+            if (@event.Questions.Count == 0)
+            {
+                ModelState.Clear();
+                ModelState.AddModelError(String.Empty, Resources.Please_specify_question);
+                ViewBag.Title = Resources.Modify_event;
+                return View("CreateEvent", @event);
+            }
+
             foreach (var question in @event.Questions)
             {
                 if (question.Answers != null)
@@ -91,22 +131,40 @@ namespace CapgeminiVoting.Controllers
                     var answerList = question.Answers.ToList();
                     answerList.RemoveAll(answer => string.IsNullOrWhiteSpace(answer.Answer));
                     question.Answers = answerList;
+
+                    if (question.Answers.Count < 2 && question.QuestionType != 2)
+                    {
+                        ModelState.Clear();
+                        ModelState.AddModelError(String.Empty, Resources.Please_specify_answer);
+                        ViewBag.Title = Resources.Modify_event;
+                        return View("CreateEvent", @event);
+                    }
                 }
             }
 
             AdminBusinessLayer.ModifyEvent(@event);
-            return RedirectToAction("Index");
+
+            ViewBag.SuccessMessage = Resources.Event_successfully_modified;
+            var indexModel = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+            return View("Index", indexModel);
         }
 
         public ActionResult EventDetails(int eventId)
         {
-            EventDetailsModel model = AdminBusinessLayer.GetEventById(eventId);
-            if (model == null)
+            if (AdminBusinessLayer.IsEventOwner(eventId, User.Identity.GetUserId()))
             {
-                return Index();
+                var model = AdminBusinessLayer.GetEventById(eventId);
+                if (model == null)
+                {
+                    return Index();
+                }
+
+                return View(model);
             }
 
-            return View(model);
+            ViewBag.ErrorMessage = Resources.Not_your_event;
+            var indexModel = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+            return View("Index", indexModel);
         }
 
         public ActionResult EventResult(int id)
@@ -117,14 +175,24 @@ namespace CapgeminiVoting.Controllers
                 return View();
             }
 
-            return View("Index");
+            ViewBag.ErrorMessage = Resources.Not_your_event;
+            var indexModel = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+            return View("Index", indexModel);
         }
 
         public ActionResult DeleteEvent(int id)
         {
-            AdminBusinessLayer.DeleteEvent(id);
-            IList<EventOverviewModel> model = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
-            return View("Index", model);
+            if (AdminBusinessLayer.IsEventOwner(id, User.Identity.GetUserId()))
+            {
+                AdminBusinessLayer.DeleteEvent(id);
+                var model = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+                ViewBag.SuccessMessage = Resources.Event_successfully_removed;
+                return View("Index", model);
+            }
+
+            ViewBag.ErrorMessage = Resources.Not_your_event;
+            var indexModel = AdminBusinessLayer.GetEventsByUser(User.Identity.GetUserId());
+            return View("Index", indexModel);
         }
 
         [HttpGet]
